requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
newspaper3k==0.2.8
google-generativeai==0.3.2
pydantic==2.6.1 
lxml==5.1.0
lxml[html_clean]==5.1.0 